+++ src/weather_data.cpp
@@ -3,7 +3,7 @@
 
 WeatherData::WeatherData(uint8_t num_hours)
 {
-  this->vals = new double[num_hours + 1];
+  this->vals = new float[num_hours + 1];
   this->minimum = 0;
   this->maxmimum = 0;
   this->mean = 0;
@@ -11,32 +11,32 @@ WeatherData::WeatherData(uint8_t num_hours)
   this->num_hours = num_hours;
 }
 
-double WeatherData::get_minimum()
+float WeatherData::get_minimum()
 {
   return this->minimum;
 }
 
-double WeatherData::get_maximum()
+float WeatherData::get_maximum()
 {
   return this->maxmimum;
 }
 
-double WeatherData::get_mean()
+float WeatherData::get_mean()
 {
   return this->mean;
 }
 
-double WeatherData::get_variance()
+float WeatherData::get_variance()
 {
   return this->variance;
 }
 
-double WeatherData::get_std()
+float WeatherData::get_std()
 {
   return sqrt(this->variance);
 }
 
-double WeatherData::get_val_at_hour(uint8_t hour)
+float WeatherData::get_val_at_hour(uint8_t hour)
 {
   return this->vals[hour];
 }
@@ -46,11 +46,11 @@ WeatherData::~WeatherData()
   delete[] this->vals;
 }
 
-void WeatherData::update_vals(double *vals)
+void WeatherData::update_vals(float *vals)
 {
-  double mean = 0.0;
-  double minimum = vals[0];
-  double maximum = vals[0];
+  float mean = 0.0;
+  float minimum = vals[0];
+  float maximum = vals[0];
   for (uint8_t i = 0; i < this->num_hours + 1; ++i)
   {
     this->vals[i] = vals[i];
@@ -72,7 +72,7 @@ void WeatherData::update_vals(double *vals)
   this->minimum = minimum;
   if (this->num_hours > 1)
   {
-    double var = 0.0;
+    float var = 0.0;
     for (int i = 1; i < this->num_hours + 1; ++i)
     {
       var += pow(vals[i] - this->mean, 2);
@@ -83,7 +83,7 @@ void WeatherData::update_vals(double *vals)
   }
 }
 
-double WeatherData::get_current()
+float WeatherData::get_current()
 {
   return this->vals[0];
 }
